refactor(frontend): split debug modal from staff pre-validate modal
Data stays centralized (one impossibility_report API response, one
backend module). The two UIs are intentionally different per the
approved brainstorm mockup — trying to share one component was forcing
either-or compromises.

PreValidationResultsModal (end-user, bunking-board SessionHeader):
- Reverts to staff-only rendering: friendly reason labels, names +
  grade only, no cm_id/gender/reason_code, no tabs, no toggle.
- Drops usePermissions dependency (modal is render-context-agnostic).
- Drops AdminReasonTable / FlatTable / JsonView and the technical
  reason map — those move to the new debug modal.

SolverDebugImpossibilityModal (new, /summer/debug/solver SweepPanel):
- Mockup-faithful header: 'Pre-validate · session=X year=Y' +
  'impossibility_report: total=N, affected_cms=M, clusters=K'.
- Three tabs: By reason (default, per-reason monospace tables with
  req_id/requester/requestee/type/detail and a technical subtitle
  per reason_code) / Flat table (sortable across
  Name/CM ID/Grade/Gender/Reason/Detail) / JSON (raw report dump).
- Cluster section lists component members with compact tuples.
- Modal is xl, scrollable, monospace; no friendly labels, no
  Ready-to-Run quick stats — pure debug surface.

SolverDebugPage now opens SolverDebugImpossibilityModal directly
(passing impossibility_report + sessionCmId + year). The
PreValidateRequestsButton on SessionHeader continues to open
PreValidationResultsModal for end users.

Test split: 6 staff-view tests (PreValidationResultsModal.test.tsx) +
13 debug-view tests (SolverDebugImpossibilityModal.test.tsx) covering
header metadata, tab rendering/default-select, By reason table /
technical subtitle / compact tuples / compact detail, Flat sortable
columns + sort reversal, JSON content, cluster member listing, and
empty state.
