+++ frontend/src/components/PreValidationResultsModal.tsx
@@ -1,4 +1,4 @@
-import { useState, useMemo } from 'react'
+import { useMemo, useState } from 'react'
 import {
   CheckCircle2,
   AlertTriangle,
@@ -17,7 +17,6 @@ import type {
   ImpossibilityReportItem,
   ImpossibilityCluster,
 } from '../services/solver'
-import { usePermissions } from '../hooks/usePermissions'
 
 interface CapacityBreakdownItem {
   campers: number
@@ -170,104 +169,18 @@ function ordinalGrade(grade: number): string {
   return `${grade}${s[(v - 20) % 10] || s[v] || s[0]} grade`
 }
 
-const TECHNICAL_REASON_DESCRIPTIONS: Record<string, string> = {
-  grade_compatibility: 'pair_grade_gap > max_range',
-  cluster_grade_compatibility: 'component grade span > max_range',
-  cross_session: 'requester.session ≠ requestee.session',
-  pair_no_shared_bunk: 'no bunk satisfies gender + grade for both campers',
-  age_pref_no_eligible_grade: 'age_preference grade-bound outside pool grades',
-  cluster_capacity: 'component_size > max_bunk_capacity',
-  malformed: 'missing requestee_id or invalid request_type',
-  target_not_in_session: 'requestee_id not in solver input',
-}
-
-function technicalReasonDescription(code: string): string {
-  return TECHNICAL_REASON_DESCRIPTIONS[code] || code
-}
-
-function shortenReqId(id: string): string {
-  if (id.length <= 11) return id
-  return `${id.slice(0, 8)}…`
-}
-
-function compactPerson(p: { name: string; cm_id: number; grade: number; gender: string }): string {
-  return `${p.name} (${p.cm_id}/g${p.grade}/${p.gender})`
-}
-
-function compactDetail(detail: Record<string, unknown> | null | undefined): string {
-  if (!detail) return ''
-  return Object.entries(detail)
-    .map(([k, v]) => `${k}=${String(v)}`)
-    .join(', ')
-}
-
-function AdminReasonTable({ items }: { items: ImpossibilityReportItem[] }) {
-  return (
-    <table className="mt-2 w-full border-collapse font-mono text-xs">
-      <thead>
-        <tr className="border-b border-amber-200 text-left text-stone-500">
-          <th className="px-1 py-1 font-normal">req_id</th>
-          <th className="px-1 py-1 font-normal">requester</th>
-          <th className="px-1 py-1 font-normal">requestee</th>
-          <th className="px-1 py-1 font-normal">type</th>
-          <th className="px-1 py-1 font-normal">detail</th>
-        </tr>
-      </thead>
-      <tbody>
-        {items.map((item) => (
-          <tr key={item.request_id} className="border-b border-amber-100/60">
-            <td className="px-1 py-1 text-stone-500">{shortenReqId(item.request_id)}</td>
-            <td className="px-1 py-1">{compactPerson(item.requester)}</td>
-            <td className="px-1 py-1">{item.requestee ? compactPerson(item.requestee) : '—'}</td>
-            <td className="px-1 py-1 text-stone-600">{item.request_type}</td>
-            <td className="px-1 py-1 text-stone-600">{compactDetail(item.detail)}</td>
-          </tr>
-        ))}
-      </tbody>
-    </table>
-  )
-}
-
-type ViewMode = 'by-reason' | 'flat' | 'json'
-
-function ImpossibilityItems({
-  items,
-  mode = 'staff',
-}: {
-  items: ImpossibilityReportItem[]
-  mode?: 'staff' | 'admin'
-}) {
+function ImpossibilityItems({ items }: { items: ImpossibilityReportItem[] }) {
   return (
     <div className="mt-2 space-y-2 border-t border-amber-200 pt-2">
       {items.map((item) => (
         <div key={item.request_id} className="text-sm">
           <div className="font-medium">
             {item.requester.name} · {ordinalGrade(item.requester.grade)}
-            {mode === 'admin' && (
-              <span className="ml-2 font-mono text-xs text-stone-500">
-                cm_id={item.requester.cm_id} · {item.requester.gender}
-              </span>
-            )}
           </div>
           {item.requestee && (
             <div className="text-stone-600">
               wants to bunk with <strong>{item.requestee.name}</strong> ·{' '}
               {ordinalGrade(item.requestee.grade)}
-              {mode === 'admin' && (
-                <span className="ml-2 font-mono text-xs text-stone-500">
-                  cm_id={item.requestee.cm_id} · {item.requestee.gender}
-                </span>
-              )}
-            </div>
-          )}
-          {mode === 'admin' && (
-            <div className="mt-2 rounded bg-stone-100 p-2 font-mono text-xs">
-              <div className="text-stone-500">{item.reason_code}</div>
-              {Object.entries(item.detail ?? {}).map(([k, v]) => (
-                <div key={k}>
-                  {k}={String(v)}
-                </div>
-              ))}
             </div>
           )}
         </div>
@@ -276,13 +189,7 @@ function ImpossibilityItems({
   )
 }
 
-function ClusterCard({
-  cluster,
-  mode,
-}: {
-  cluster: ImpossibilityCluster
-  mode: 'staff' | 'admin'
-}) {
+function ClusterCard({ cluster }: { cluster: ImpossibilityCluster }) {
   return (
     <div className="rounded-lg border border-amber-200 bg-amber-50 p-3">
       <div className="font-semibold text-amber-900">{friendlyReasonLabel(cluster.reason_code)}</div>
@@ -293,137 +200,11 @@ function ClusterCard({
             <div key={c.cm_id} className="flex flex-wrap items-center gap-x-2">
               <span className="font-medium">{c.name}</span>
               <span className="text-stone-600">· {ordinalGrade(c.grade)}</span>
-              {mode === 'admin' && (
-                <span className="font-mono text-xs text-stone-500">
-                  cm_id={c.cm_id} · {c.gender}
-                </span>
-              )}
-            </div>
-          ))}
-        </div>
-      )}
-      {mode === 'admin' && (
-        <div className="mt-2 rounded bg-stone-100 p-2 font-mono text-xs">
-          <div className="text-stone-500">{cluster.reason_code}</div>
-          {Object.entries(cluster.detail ?? {}).map(([k, v]) => (
-            <div key={k}>
-              {k}={String(v)}
             </div>
           ))}
         </div>
       )}
     </div>
-  )
-}
-
-type FlatTableSortColumn = 'name' | 'cm_id' | 'grade' | 'gender' | 'reason'
-type FlatTableSortDir = 'asc' | 'desc'
-
-function FlatTable({ items }: { items: ImpossibilityReportItem[] }) {
-  const [sortCol, setSortCol] = useState<FlatTableSortColumn>('name')
-  const [sortDir, setSortDir] = useState<FlatTableSortDir>('asc')
-
-  const sortedItems = useMemo(() => {
-    const arr = [...items]
-    arr.sort((a, b) => {
-      let av: string | number
-      let bv: string | number
-      switch (sortCol) {
-        case 'name':
-          av = a.requester.name
-          bv = b.requester.name
-          break
-        case 'cm_id':
-          av = a.requester.cm_id
-          bv = b.requester.cm_id
-          break
-        case 'grade':
-          av = a.requester.grade
-          bv = b.requester.grade
-          break
-        case 'gender':
-          av = a.requester.gender
-          bv = b.requester.gender
-          break
-        case 'reason':
-          av = a.reason_code
-          bv = b.reason_code
-          break
-      }
-      if (av < bv) return sortDir === 'asc' ? -1 : 1
-      if (av > bv) return sortDir === 'asc' ? 1 : -1
-      return 0
-    })
-    return arr
-  }, [items, sortCol, sortDir])
-
-  const handleSort = (col: FlatTableSortColumn) => {
-    if (sortCol === col) {
-      setSortDir((d) => (d === 'asc' ? 'desc' : 'asc'))
-    } else {
-      setSortCol(col)
-      setSortDir('asc')
-    }
-  }
-
-  const arrow = (col: FlatTableSortColumn) =>
-    sortCol === col ? (sortDir === 'asc' ? ' ↑' : ' ↓') : ''
-
-  const headerCell = (col: FlatTableSortColumn, label: string) => (
-    <th
-      scope="col"
-      onClick={() => handleSort(col)}
-      className="cursor-pointer border-b border-stone-300 px-2 py-1 text-left font-semibold select-none hover:bg-stone-200"
-    >
-      {label}
-      {arrow(col)}
-    </th>
-  )
-
-  return (
-    <div className="overflow-x-auto">
-      <table className="w-full border-collapse text-xs">
-        <thead className="bg-stone-100">
-          <tr>
-            {headerCell('name', 'Name')}
-            {headerCell('cm_id', 'CM ID')}
-            {headerCell('grade', 'Grade')}
-            {headerCell('gender', 'Gender')}
-            {headerCell('reason', 'Reason')}
-            <th scope="col" className="border-b border-stone-300 px-2 py-1 text-left font-semibold">
-              Detail
-            </th>
-          </tr>
-        </thead>
-        <tbody>
-          {sortedItems.map((item) => (
-            <tr key={item.request_id} className="border-b border-stone-200">
-              <td className="px-2 py-1">{item.requester.name}</td>
-              <td className="px-2 py-1 font-mono">{item.requester.cm_id}</td>
-              <td className="px-2 py-1">{item.requester.grade}</td>
-              <td className="px-2 py-1">{item.requester.gender}</td>
-              <td className="px-2 py-1 font-mono">{item.reason_code}</td>
-              <td className="px-2 py-1 font-mono text-stone-600">
-                {Object.entries(item.detail ?? {})
-                  .map(([k, v]) => `${k}=${String(v)}`)
-                  .join(', ')}
-              </td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
-  )
-}
-
-function JsonView({ report }: { report: ImpossibilityReport }) {
-  return (
-    <pre
-      data-testid="impossibility-json"
-      className="max-h-96 overflow-auto rounded bg-stone-100 p-3 font-mono text-xs"
-    >
-      {JSON.stringify(report, null, 2)}
-    </pre>
   )
 }
 
@@ -489,8 +270,6 @@ export default function PreValidationResultsModal({
   results,
 }: PreValidationResultsModalProps) {
   const [showDetails, setShowDetails] = useState(false)
-  const { isAdmin } = usePermissions()
-  const [viewMode, setViewMode] = useState<ViewMode>('by-reason')
 
   const { valid, errors, warnings, statistics, impossibility_report } = results
 
@@ -561,7 +340,7 @@ export default function PreValidationResultsModal({
       onClose={onClose}
       header={headerContent}
       footer={footerContent}
-      size={isAdmin ? 'xl' : 'lg'}
+      size="lg"
       noPadding
       scrollable
     >
@@ -660,84 +439,34 @@ export default function PreValidationResultsModal({
         </div>
       )}
 
-      {/* Impossibility Report */}
+      {/* Impossibility Report — staff view */}
       <div className="space-y-3 px-5 py-4">
-        {isAdmin &&
-          (impossibility_report.total_impossible > 0 ||
-            impossibility_report.clusters.length > 0) && (
-            <div role="tablist" className="mb-2 flex gap-1 border-b border-stone-200">
-              {(['by-reason', 'flat', 'json'] as const).map((tab) => {
-                const label =
-                  tab === 'by-reason' ? 'By reason' : tab === 'flat' ? 'Flat table' : 'JSON'
-                const active = viewMode === tab
-                return (
-                  <button
-                    key={tab}
-                    role="tab"
-                    aria-selected={active}
-                    onClick={() => setViewMode(tab)}
-                    className={`-mb-px border-b-2 px-3 py-1.5 text-sm transition-colors ${
-                      active
-                        ? 'border-forest-500 text-forest-700 font-medium'
-                        : 'border-transparent text-stone-600 hover:text-stone-900'
-                    }`}
-                  >
-                    {label}
-                  </button>
-                )
-              })}
-            </div>
-          )}
-
         {impossibility_report.total_impossible === 0 &&
         impossibility_report.clusters.length === 0 ? (
           <div className="rounded-lg border border-green-200 bg-green-50 p-4 text-green-800">
             <CheckCircle2 className="mr-2 inline-block h-5 w-5" />
             No impossible requests found for this scenario.
           </div>
-        ) : isAdmin && viewMode === 'flat' ? (
-          <FlatTable items={impossibility_report.flat} />
-        ) : isAdmin && viewMode === 'json' ? (
-          <JsonView report={impossibility_report} />
         ) : (
           <>
-            {Object.entries(impossibility_report.by_reason).map(([code, items]) =>
-              isAdmin ? (
-                <details
-                  key={code}
-                  open
-                  className="rounded-md border border-amber-200 bg-amber-50 p-3 font-mono"
-                >
-                  <summary className="flex cursor-pointer items-center justify-between font-semibold text-amber-900">
-                    <span>{code}</span>
-                    <span className="rounded-full bg-amber-400 px-2 py-0.5 text-[10px] font-bold text-white">
-                      {items.length}
-                    </span>
-                  </summary>
-                  <div className="mt-1 text-[11px] text-stone-600">
-                    {technicalReasonDescription(code)}
-                  </div>
-                  <AdminReasonTable items={items} />
-                </details>
-              ) : (
-                <details
-                  key={code}
-                  open
-                  className="rounded-lg border border-amber-200 bg-amber-50 p-3"
-                >
-                  <summary className="flex cursor-pointer items-center justify-between font-semibold text-amber-900">
-                    <span>{friendlyReasonLabel(code)}</span>
-                    <span className="rounded-full bg-amber-400 px-2 py-0.5 text-xs font-bold text-white">
-                      {items.length}
-                    </span>
-                  </summary>
-                  <ImpossibilityItems items={items} mode="staff" />
-                </details>
-              )
-            )}
+            {Object.entries(impossibility_report.by_reason).map(([code, items]) => (
+              <details
+                key={code}
+                open
+                className="rounded-lg border border-amber-200 bg-amber-50 p-3"
+              >
+                <summary className="flex cursor-pointer items-center justify-between font-semibold text-amber-900">
+                  <span>{friendlyReasonLabel(code)}</span>
+                  <span className="rounded-full bg-amber-400 px-2 py-0.5 text-xs font-bold text-white">
+                    {items.length}
+                  </span>
+                </summary>
+                <ImpossibilityItems items={items} />
+              </details>
+            ))}
 
             {impossibility_report.clusters.map((cluster, idx) => (
-              <ClusterCard key={idx} cluster={cluster} mode={isAdmin ? 'admin' : 'staff'} />
+              <ClusterCard key={idx} cluster={cluster} />
             ))}
           </>
         )}
